feat: move kubernetes_configmap_jwt_scan's config to top level script
